Refactor(gamePlayController): mainController의 play 역할을 gamePlayController 가 하게 함
