+++ src/main/java/baseball/controller/MainController.java
@@ -10,16 +10,13 @@ import baseball.view.OutputView;
 import java.util.List;
 
 public class MainController {
-    public static MainController instance = new MainController(InputView.getInstance(), OutputView.getInstance(), GameRestartController.getInstance());
-    private final InputView inputView;
-    private final OutputView outputView;
+    public static MainController instance = new MainController(GameRestartController.getInstance(), GamePlayController.getInstance());
     private final GameRestartController gameRestartController;
-    private GameManager gameManager;
+    private final GamePlayController gamePlayController;
 
-    private MainController(InputView inputView, OutputView outputView, GameRestartController gameRestartController) {
-        this.inputView = inputView;
-        this.outputView = outputView;
+    private MainController(GameRestartController gameRestartController, GamePlayController gamePlayController) {
         this.gameRestartController = gameRestartController;
+        this.gamePlayController = gamePlayController;
     }
 
     public static MainController getInstance() {
@@ -29,40 +26,10 @@ public class MainController {
     public void run() {
         boolean continuePlaying = true;
         while (continuePlaying) {
-            continuePlaying = play();
+            continuePlaying = gamePlayController.play();
+            if (continuePlaying) {
+                continuePlaying = gameRestartController.wantsRestart();
+            }
         }
-    }
-
-    private boolean play() {
-        outputView.printStart();
-        initializeGameManager();
-        boolean needsNextRound = true;
-        while (needsNextRound) {
-            UserNumber userNumber = readUserNumber();
-            needsNextRound = playRound(userNumber);
-            printRoundResult();
-        }
-        return gameRestartController.wantsRestart();
-    }
-
-    private void initializeGameManager() {
-        ComputerNumber computerNumber = ComputerNumber.create();
-        gameManager = GameManager.from(computerNumber);
-    }
-
-    private UserNumber readUserNumber() {
-        List<Integer> numbers = inputView.readUserNumber();
-        return UserNumber.from(numbers);
-    }
-
-    private boolean playRound(UserNumber userNumber) {
-        boolean needsNextRound;
-        needsNextRound = gameManager.playRound(userNumber);
-        return needsNextRound;
-    }
-
-    private void printRoundResult() {
-        RoundResultDto resultDto = gameManager.createResultDto();
-        outputView.printResult(resultDto.isNothing(), resultDto.getBallCount(), resultDto.getStrikeCount());
     }
 }
